+++ src/main/java/com/life/majiang/community/community/dto/QuestionDTO.java
@@ -5,13 +5,13 @@ import lombok.Data;
 
 @Data
 public class QuestionDTO {
-    private int id;
+    private long id;
     private String title;
     private String description;
     private String tag;
     private long gmtCreate;
     private long gmtModify;
-    private int creator;
+    private long creator;
     private int viewCount;
     private int commentCount;
     private int likeCount;
